requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
pytest==8.4.1
# Required by FastAPI/Starlette TestClient during pytest runs.
httpx==0.28.1
chromadb==1.0.8
pypdf==5.4.0
python-docx==1.1.2
